split test_utility.py into util-specific tests
